requirements.txt
absl-py==1.0.0
argon2-cffi==21.1.0
asgiref==3.4.1
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
autocorrect==2.6.1
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
bleach==4.1.0
bs4==0.0.1
cachetools==4.2.4
cairocffi==1.3.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
ctc==2.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.4
Django==4.0
docopt==0.6.2
docutils==0.15.2
easyocr==1.4.2
editdistance==0.6.0
emoji==1.7.0
englisttohindi==4.1.0
entrypoints==0.3
filelock==3.4.2
Flask==2.0.2
Flask-Cors==3.0.10
flatbuffers==2.0
fonttools==4.28.3
func-timeout==4.3.5
gast==0.4.0
google-api-core==2.11.0
google-api-python-client==2.81.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.58.0
googlesearch-python==1.1.0
grpcio==1.43.0
gunicorn==20.1.0
h11==0.13.0
h5py==3.6.0
httplib2==0.21.0
ibm-cos-sdk==2.7.0
ibm-cos-sdk-core==2.7.0
ibm-cos-sdk-s3transfer==2.7.0
ibm-watson-machine-learning==1.0.286
idna==2.10
imageio==2.16.2
importlib-metadata==4.10.0
importlib-resources==5.4.0
imutils==0.5.4
indicate==0.1.0
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
lmdb==1.3.0
lomond==0.3.3
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mistune==0.8.4
mlxtend==0.19.0
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.8
nltk==3.7
notebook==6.4.11
numpy==1.21.4
oauthlib==3.1.1
opencv-python==4.5.5.62
opencv-python-headless==4.5.4.60
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.3
path==16.4.0
pickleshare==0.7.5
Pillow==8.4.0
pipreqs==0.4.11
platformdirs==2.4.1
prometheus-client==0.12.0
prompt-toolkit==3.0.23
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.2
pydot-ng==2.0.0
Pygments==2.10.0
pyOpenSSL==22.0.0
pyparsing==3.0.6
pyrsistent==0.18.0
pytesseract==0.3.8
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.3.0
pywin32==302
pywinpty==1.1.6
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.1
QtPy==1.11.3
regex==2022.3.15
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
selenium==4.1.0
Send2Trash==1.8.0
setuptools-scm==6.3.2
Shapely==1.8.0
six==1.12.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2
sqlparse==0.4.2
tabulate==0.9.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-addons==0.17.1
tensorflow-estimator==2.7.0
tensorflow-gpu==2.8.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
terminado==0.12.1
tesseract==0.1.3
testpath==0.5.0
textblob==0.17.1
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.0.0
tifffile==2022.4.8
tomli==1.2.2
torch==1.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.63.1
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
typeguard==2.13.3
typing_extensions==4.0.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.13.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
whois==0.9.27
widgetsnbextension==3.5.2
wrapt==1.13.3
wsproto==1.0.0
xgboost==1.6.1
yarg==0.1.9
zipp==3.6.0
